AccountID型導入

+++ ddl/ddl.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `account` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `id` int NOT NULL AUTO_INCREMENT,
   `username` varchar(255) NOT NULL UNIQUE,
   `password_hash` varchar(255) NOT NULL,
   `display_name` varchar(255),
@@ -11,8 +11,8 @@ CREATE TABLE `account` (
 );
 
 CREATE TABLE `status` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `account_id` bigint(20) NOT NULL,
+  `id` int NOT NULL AUTO_INCREMENT,
+  `account_id` int NOT NULL,
   `url` varchar(255),
   `content` text NOT NULL,
   `create_at` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP,
